requirements.txt
dj-database-url==0.5.0
Django==2.2.3
django-heroku==0.3.1
djangorestframework==3.10.2
gunicorn==19.9.0
pkg-resources==0.0.0
psycopg2==2.8.3
pyparsing==2.4.2
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
whitenoise==4.1.4
